backend/requirements.txt
asgiref==3.5.2
astroid==2.12.11
async-generator==1.10
attrs==22.1.0
black==22.10.0
boto3==1.24.89
botocore==1.27.89
certifi==2022.9.24
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
dill==0.3.5.1
distlib==0.3.6
Django==4.1.1
django-cors-headers==3.13.0
django-filter
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
exceptiongroup==1.0.0rc9
filelock==3.8.0
flake8==5.0.4
h11==0.14.0
humanize
identify==2.5.6
idna==3.4
isort==5.10.1
isort
jmespath==1.0.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
nodeenv==1.7.0
outcome==1.2.0
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pre-commit==2.20.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.5.0
pylint==2.15.4
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
pytz
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.12
virtualenv==20.16.5
webdriver-manager==3.8.3
wrapt==1.14.1
wsproto==1.2.0
mysqlclient